+++ cpp/src/ffi/reader_c.cpp
@@ -466,138 +466,6 @@ LoonFFIResult loon_get_record_batch_reader(LoonReaderHandle reader,
   RETURN_UNREACHABLE();
 }
 
-// ==================== Per-batch RecordBatchReader API ====================
-//
-// ArrowArrayStream-based iteration (`loon_get_record_batch_reader` above)
-// turned out to be incompatible with the Arrow Java `ArrowReader`'s
-// shared-VectorSchemaRoot pattern: when the underlying reader emits
-// `RecordBatch::Slice` results whose ArrowArray carries a non-zero
-// `offset`, the JVM side ignores the offset and keeps reading from buffer
-// position 0 — every batch after the first shows the same data. Milvus's
-// Go binding (`internal/core/src/segcore/packed_reader_c.cpp ReadNext`)
-// sidesteps this by exporting a fresh ArrowArray+ArrowSchema pair per
-// batch. The functions below provide the same shape for the Scala/JNI
-// path so the Spark connector can mirror Milvus's proven scheme.
-//
-// Lifecycle:
-//   - `loon_record_batch_reader_new` opens a handle, owning a
-//     `shared_ptr<arrow::RecordBatchReader>`.
-//   - `loon_record_batch_reader_read_next` fills caller-owned
-//     ArrowArray/ArrowSchema structs (allocated + zero-initialized by the
-//     caller) for each batch; on EOF the `release` fields are left NULL.
-//   - `loon_record_batch_reader_destroy` drops the handle.
-
-struct RecordBatchReaderHolder {
-  std::shared_ptr<arrow::RecordBatchReader> reader;
-};
-
-LoonFFIResult loon_record_batch_reader_new(LoonReaderHandle reader,
-                                           const char* predicate,
-                                           LoonRecordBatchReaderHandle* out_handle) {
-  if (!reader || !out_handle) {
-    RETURN_ERROR(LOON_INVALID_ARGS, "Invalid arguments: reader and out_handle must not be null");
-  }
-
-  try {
-    auto* cpp_reader = reinterpret_cast<Reader*>(reader);
-    std::string predicate_str = predicate ? predicate : "";
-
-    auto result = cpp_reader->get_record_batch_reader(predicate_str);
-    if (!result.ok()) {
-      RETURN_ERROR(LOON_ARROW_ERROR, result.status().ToString());
-    }
-
-    auto* holder = new RecordBatchReaderHolder{result.ValueOrDie()};
-    *out_handle = reinterpret_cast<LoonRecordBatchReaderHandle>(holder);
-    RETURN_SUCCESS();
-  } catch (std::exception& e) {
-    RETURN_EXCEPTION(e.what());
-  }
-
-  RETURN_UNREACHABLE();
-}
-
-LoonFFIResult loon_record_batch_reader_read_next(LoonRecordBatchReaderHandle handle,
-                                                 struct ArrowArray* out_array,
-                                                 struct ArrowSchema* out_schema) {
-  if (!handle || !out_array || !out_schema) {
-    RETURN_ERROR(LOON_INVALID_ARGS, "Invalid arguments: handle, out_array, out_schema must not be null");
-  }
-
-  try {
-    auto* holder = reinterpret_cast<RecordBatchReaderHolder*>(handle);
-    std::shared_ptr<arrow::RecordBatch> batch;
-    auto status = holder->reader->ReadNext(&batch);
-    if (!status.ok()) {
-      RETURN_ERROR(LOON_ARROW_ERROR, status.ToString());
-    }
-
-    // PackedRecordBatchReader::ReadNext can hand back a RecordBatch whose
-    // column arrays carry a non-zero `offset` — this happens whenever the
-    // underlying chunk is larger than min_rows and the remainder is kept
-    // in the queue via `rb->Slice(min_rows)` (see reader.cpp). ArrowArray's
-    // C Data Interface specifies consumers must honour `offset`, but Arrow
-    // Java's `Data.importVectorSchemaRoot` ignores it. Exporting a sliced
-    // batch to a Java consumer therefore produces a row-cycle: every
-    // post-slice batch re-reads from buffer position 0, duplicating the
-    // first min_rows rows' data.
-    //
-    // Materialize sliced columns into fresh offset=0 arrays via
-    // arrow::Concatenate (allocates new buffers, copies only the slice
-    // range). Non-sliced columns pass through unchanged.
-    if (batch != nullptr) {
-      bool has_sliced_column = false;
-      for (int i = 0; i < batch->num_columns(); ++i) {
-        if (batch->column(i)->offset() != 0) {
-          has_sliced_column = true;
-          break;
-        }
-      }
-      if (has_sliced_column) {
-        std::vector<std::shared_ptr<arrow::Array>> fresh_cols;
-        fresh_cols.reserve(batch->num_columns());
-        for (int i = 0; i < batch->num_columns(); ++i) {
-          auto col = batch->column(i);
-          if (col->offset() == 0) {
-            fresh_cols.push_back(col);
-          } else {
-            auto concat_result = arrow::Concatenate({col}, arrow::default_memory_pool());
-            if (!concat_result.ok()) {
-              RETURN_ERROR(LOON_ARROW_ERROR, concat_result.status().ToString());
-            }
-            fresh_cols.push_back(concat_result.ValueOrDie());
-          }
-        }
-        batch = arrow::RecordBatch::Make(batch->schema(), batch->num_rows(), fresh_cols);
-      }
-    }
-
-    if (batch == nullptr) {
-      // EOF: leave release callbacks NULL so the caller can detect it.
-      out_array->release = nullptr;
-      out_schema->release = nullptr;
-      RETURN_SUCCESS();
-    }
-
-    auto export_status = arrow::ExportRecordBatch(*batch, out_array, out_schema);
-    if (!export_status.ok()) {
-      RETURN_ERROR(LOON_ARROW_ERROR, export_status.ToString());
-    }
-
-    RETURN_SUCCESS();
-  } catch (std::exception& e) {
-    RETURN_EXCEPTION(e.what());
-  }
-
-  RETURN_UNREACHABLE();
-}
-
-void loon_record_batch_reader_destroy(LoonRecordBatchReaderHandle handle) {
-  if (!handle)
-    return;
-  delete reinterpret_cast<RecordBatchReaderHolder*>(handle);
-}
-
 LoonFFIResult loon_get_chunk_reader(LoonReaderHandle reader,
                                     int64_t column_group_id,
                                     const char* const* needed_columns,
